Revert "Weighted from int to double weights"
This reverts commit 7570ac2ffee1b56ee55567da8c255a37189d2a3f.

+++ src/wisard_bind.cc
@@ -163,7 +163,7 @@ PYBIND11_MODULE(wisardpkg, m){
     ;    
     
     py::class_<Weighted, ClassificationBase>(m, "Weighted", py::module_local())
-      .def(py::init<std::map<std::string,std::vector<double>>>())
+      .def(py::init<std::map<std::string,std::vector<int>>>())
       .def("setWeights", &Weighted::setWeights)
       .def("getWeights", &Weighted::getWeights)
     ;
